refactor: remove dead C++ expression layer (expr/, graph/)
The C++ expression AST (Expr.hpp, ExprRewrite, ExprStructural, ExprEval)
and the GraphTypes.hpp alias hub were not used by the JIT, runtime, or
C API. The JIT works directly from the FlatProgram instruction stream
produced by the TS emitter — it never touches the C++ expression tree.

The only live dependency was ControlParam (lock-free atomic parameter
struct), extracted into engine/ControlParam.hpp (~25 lines) to replace
the 825-line Expr.hpp.

Removed 2,674 lines of C++ and 2 compiled translation units from the
shared library.
